+++ packages/block-library/src/accordion-heading/index.js
@@ -1,11 +1,15 @@
 /**
+ * WordPress dependencies
+ */
+import { accordionHeading as icon } from '@wordpress/icons';
+
+/**
  * Internal dependencies
  */
 import edit from './edit';
 import save from './save';
 import metadata from './block.json';
 import initBlock from '../utils/init-block';
-import icon from './icon';
 import deprecated from './deprecated';
 
 const { name } = metadata;