+++ packages/core-data/src/selectors.ts
@@ -52,9 +52,16 @@ export interface State {
 	registeredPostMeta: Record< string, Object >;
 	editorSettings: Record< string, any > | null;
 	editorAssets: Record< string, any > | null;
+	icons: Icon[];
 }
 
 type EntityRecordKey = string | number;
+
+export interface Icon {
+	name: string;
+	content: string;
+	label: string;
+}
 
 interface EntitiesState {
 	config: EntityConfig[];